Moved cellEFP to a different directory for clarity
